# Fix the perf issue in TensorPrimitives.MinMaxCore