--- a/src/libraries/System.Numerics.Tensors/src/System/Numerics/Tensors/TensorPrimitives.cs
+++ b/src/libraries/System.Numerics.Tensors/src/System/Numerics/Tensors/TensorPrimitives.cs
@@ -169,20 +169,8 @@ namespace System.Numerics.Tensors
         /// operating systems or architectures.
         /// </para>
         /// </remarks>
-        public static float CosineSimilarity(ReadOnlySpan<float> x, ReadOnlySpan<float> y)
-        {
-            if (x.IsEmpty)
-            {
-                ThrowHelper.ThrowArgument_SpansMustBeNonEmpty();
-            }
-
-            if (x.Length != y.Length)
-            {
-                ThrowHelper.ThrowArgument_SpansMustHaveSameLength();
-            }
-
-            return CosineSimilarityCore(x, y);
-        }
+        public static float CosineSimilarity(ReadOnlySpan<float> x, ReadOnlySpan<float> y) =>
+            CosineSimilarityCore(x, y);
 
         /// <summary>Computes the distance between two points, specified as non-empty, equal-length tensors of single-precision floating-point numbers, in Euclidean space.</summary>
         /// <param name="x">The first tensor, represented as a span.</param>
@@ -314,15 +302,8 @@ namespace System.Numerics.Tensors
         /// operating systems or architectures.
         /// </para>
         /// </remarks>
-        public static int IndexOfMax(ReadOnlySpan<float> x)
-        {
-            if (x.IsEmpty)
-            {
-                return -1;
-            }
-
-            return IndexOfMinMaxCore<IndexOfMaxOperator>(x);
-        }
+        public static int IndexOfMax(ReadOnlySpan<float> x) =>
+            IndexOfMinMaxCore<IndexOfMaxOperator>(x);
 
         /// <summary>Searches for the index of the single-precision floating-point number with the largest magnitude in the specified tensor.</summary>
         /// <param name="x">The tensor, represented as a span.</param>
@@ -338,15 +319,8 @@ namespace System.Numerics.Tensors
         /// operating systems or architectures.
         /// </para>
         /// </remarks>
-        public static int IndexOfMaxMagnitude(ReadOnlySpan<float> x)
-        {
-            if (x.IsEmpty)
-            {
-                return -1;
-            }
-
-            return IndexOfMinMaxCore<IndexOfMaxMagnitudeOperator>(x);
-        }
+        public static int IndexOfMaxMagnitude(ReadOnlySpan<float> x) =>
+            IndexOfMinMaxCore<IndexOfMaxMagnitudeOperator>(x);
 
         /// <summary>Searches for the index of the smallest single-precision floating-point number in the specified tensor.</summary>
         /// <param name="x">The tensor, represented as a span.</param>
@@ -361,15 +335,8 @@ namespace System.Numerics.Tensors
         /// operating systems or architectures.
         /// </para>
         /// </remarks>
-        public static int IndexOfMin(ReadOnlySpan<float> x)
-        {
-            if (x.IsEmpty)
-            {
-                return -1;
-            }
-
-            return IndexOfMinMaxCore<IndexOfMinOperator>(x);
-        }
+        public static int IndexOfMin(ReadOnlySpan<float> x) =>
+            IndexOfMinMaxCore<IndexOfMinOperator>(x);
 
         /// <summary>Searches for the index of the single-precision floating-point number with the smallest magnitude in the specified tensor.</summary>
         /// <param name="x">The tensor, represented as a span.</param>
@@ -385,15 +352,8 @@ namespace System.Numerics.Tensors
         /// operating systems or architectures.
         /// </para>
         /// </remarks>
-        public static int IndexOfMinMagnitude(ReadOnlySpan<float> x)
-        {
-            if (x.IsEmpty)
-            {
-                return -1;
-            }
-
-            return IndexOfMinMaxCore<IndexOfMinMagnitudeOperator>(x);
-        }
+        public static int IndexOfMinMagnitude(ReadOnlySpan<float> x) =>
+            IndexOfMinMaxCore<IndexOfMinMagnitudeOperator>(x);
 
         /// <summary>Computes the element-wise natural (base <c>e</c>) logarithm of single-precision floating-point numbers in the specified tensor.</summary>
         /// <param name="x">The tensor, represented as a span.</param>
@@ -455,15 +415,8 @@ namespace System.Numerics.Tensors
         /// operating systems or architectures.
         /// </para>
         /// </remarks>
-        public static float Max(ReadOnlySpan<float> x)
-        {
-            if (x.IsEmpty)
-            {
-                ThrowHelper.ThrowArgument_SpansMustBeNonEmpty();
-            }
-
-            return MinMaxCore<MaxOperator>(x);
-        }
+        public static float Max(ReadOnlySpan<float> x) =>
+            MinMaxCore<MaxOperator>(x);
 
         /// <summary>Computes the element-wise maximum of the single-precision floating-point numbers in the specified tensors.</summary>
         /// <param name="x">The first tensor, represented as a span.</param>
@@ -504,15 +457,8 @@ namespace System.Numerics.Tensors
         /// operating systems or architectures.
         /// </para>
         /// </remarks>
-        public static float MaxMagnitude(ReadOnlySpan<float> x)
-        {
-            if (x.IsEmpty)
-            {
-                ThrowHelper.ThrowArgument_SpansMustBeNonEmpty();
-            }
-
-            return MinMaxCore<MaxMagnitudeOperator>(x);
-        }
+        public static float MaxMagnitude(ReadOnlySpan<float> x) =>
+            MinMaxCore<MaxMagnitudeOperator>(x);
 
         /// <summary>Computes the element-wise single-precision floating-point number with the largest magnitude in the specified tensors.</summary>
         /// <param name="x">The first tensor, represented as a span.</param>
@@ -546,15 +492,8 @@ namespace System.Numerics.Tensors
         /// operating systems or architectures.
         /// </para>
         /// </remarks>
-        public static float Min(ReadOnlySpan<float> x)
-        {
-            if (x.IsEmpty)
-            {
-                ThrowHelper.ThrowArgument_SpansMustBeNonEmpty();
-            }
-
-            return MinMaxCore<MinOperator>(x);
-        }
+        public static float Min(ReadOnlySpan<float> x) =>
+            MinMaxCore<MinOperator>(x);
 
         /// <summary>Computes the element-wise minimum of the single-precision floating-point numbers in the specified tensors.</summary>
         /// <param name="x">The first tensor, represented as a span.</param>
@@ -595,15 +534,8 @@ namespace System.Numerics.Tensors
         /// operating systems or architectures.
         /// </para>
         /// </remarks>
-        public static float MinMagnitude(ReadOnlySpan<float> x)
-        {
-            if (x.IsEmpty)
-            {
-                ThrowHelper.ThrowArgument_SpansMustBeNonEmpty();
-            }
-
-            return MinMaxCore<MinMagnitudeOperator>(x);
-        }
+        public static float MinMagnitude(ReadOnlySpan<float> x) =>
+            MinMaxCore<MinMagnitudeOperator>(x);
 
         /// <summary>Computes the element-wise single-precision floating-point number with the smallest magnitude in the specified tensors.</summary>
         /// <param name="x">The first tensor, represented as a span.</param>