This resolves a performance issue caused by `GetFirstNaN` and it not being inlined. The perf is now on par or better than `IndexOfMax`, as expected.

| Method            | TensorLength | Mean          | Error       | StdDev      | Median        |
|------------------ |------------- |--------------:|------------:|------------:|--------------:|
| Max (main)        | 1            |      2.478 ns |   0.0672 ns |   0.0800 ns |      2.416 ns |
| Max (pr)          | 1            |      1.636 ns |   0.0199 ns |   0.0166 ns |      1.638 ns |
| Max (main)        | 2            |      2.994 ns |   0.0353 ns |   0.0330 ns |      2.986 ns |
| Max (pr)          | 2            |      2.203 ns |   0.0472 ns |   0.0442 ns |      2.196 ns |
| Max (main)        | 3            |      3.524 ns |   0.0446 ns |   0.0417 ns |      3.513 ns |
| Max (pr)          | 3            |      2.576 ns |   0.0526 ns |   0.0492 ns |      2.538 ns |
| Max (main)        | 4            |      3.451 ns |   0.0579 ns |   0.0541 ns |      3.451 ns |
| Max (pr)          | 4            |      2.220 ns |   0.0547 ns |   0.0512 ns |      2.208 ns |
| Max (main)        | 5            |      5.034 ns |   0.0723 ns |   0.0676 ns |      4.989 ns |
| Max (pr)          | 5            |      2.940 ns |   0.0521 ns |   0.0487 ns |      2.921 ns |
| Max (main)        | 6            |      5.069 ns |   0.0896 ns |   0.0795 ns |      5.066 ns |
| Max (pr)          | 6            |      2.962 ns |   0.0544 ns |   0.0509 ns |      2.959 ns |
| Max (main)        | 7            |      5.122 ns |   0.0851 ns |   0.0754 ns |      5.134 ns |
| Max (pr)          | 7            |      2.929 ns |   0.0278 ns |   0.0232 ns |      2.931 ns |
| Max (main)        | 8            |      4.162 ns |   0.0568 ns |   0.0532 ns |      4.164 ns |
| Max (pr)          | 8            |      2.386 ns |   0.0546 ns |   0.0511 ns |      2.372 ns |
| Max (main)        | 9            |      6.131 ns |   0.0758 ns |   0.0709 ns |      6.135 ns |
| Max (pr)          | 9            |      3.166 ns |   0.0697 ns |   0.0652 ns |      3.150 ns |
| Max (main)        | 10           |      6.098 ns |   0.0670 ns |   0.0626 ns |      6.093 ns |
| Max (pr)          | 10           |      3.165 ns |   0.0499 ns |   0.0467 ns |      3.179 ns |
| Max (main)        | 11           |      5.720 ns |   0.0635 ns |   0.0594 ns |      5.690 ns |
| Max (pr)          | 11           |      3.104 ns |   0.0320 ns |   0.0284 ns |      3.087 ns |
| Max (main)        | 12           |      5.692 ns |   0.0633 ns |   0.0592 ns |      5.687 ns |
| Max (pr)          | 12           |      3.147 ns |   0.0698 ns |   0.0653 ns |      3.147 ns |
| Max (main)        | 13           |      5.643 ns |   0.0249 ns |   0.0194 ns |      5.639 ns |
| Max (pr)          | 13           |      3.171 ns |   0.0798 ns |   0.0747 ns |      3.144 ns |
| Max (main)        | 14           |      5.694 ns |   0.0933 ns |   0.0827 ns |      5.672 ns |
| Max (pr)          | 14           |      3.111 ns |   0.0465 ns |   0.0412 ns |      3.085 ns |
| Max (main)        | 15           |      6.223 ns |   0.0762 ns |   0.0713 ns |      6.232 ns |
| Max (pr)          | 15           |      3.102 ns |   0.0401 ns |   0.0355 ns |      3.085 ns |
| Max (main)        | 16           |      4.541 ns |   0.0671 ns |   0.0627 ns |      4.553 ns |
| Max (pr)          | 16           |      2.584 ns |   0.0625 ns |   0.0584 ns |      2.579 ns |
| Max (main)        | 32           |      7.656 ns |   0.1455 ns |   0.1361 ns |      7.682 ns |
| Max (pr)          | 32           |      3.725 ns |   0.0708 ns |   0.0662 ns |      3.730 ns |
| Max (main)        | 64           |     39.188 ns |   0.4249 ns |   0.3975 ns |     38.957 ns |
| Max (pr)          | 64           |      7.028 ns |   0.1278 ns |   0.1196 ns |      6.992 ns |
| Max (main)        | 128          |     26.486 ns |   0.4680 ns |   0.4378 ns |     26.348 ns |
| Max (pr)          | 128          |     15.763 ns |   0.2658 ns |   0.2486 ns |     15.842 ns |
| Max (main)        | 256          |    131.064 ns |   1.6986 ns |   1.5889 ns |    130.746 ns |
| Max (pr)          | 256          |     31.814 ns |   0.5802 ns |   0.5427 ns |     31.491 ns |
| Max (main)        | 512          |    261.370 ns |   3.2722 ns |   3.0608 ns |    261.329 ns |
| Max (pr)          | 512          |     71.107 ns |   1.3994 ns |   1.3090 ns |     70.427 ns |
| Max (main)        | 1024         |    510.065 ns |   5.4908 ns |   5.1361 ns |    509.669 ns |
| Max (pr)          | 1024         |    150.685 ns |   2.8876 ns |   3.2096 ns |    148.938 ns |
| Max (main)        | 2048         |    993.753 ns |  11.4559 ns |  10.7159 ns |    992.964 ns |
| Max (pr)          | 2048         |    304.270 ns |   5.4914 ns |   5.1366 ns |    300.003 ns |
| Max (main)        | 4096         |  1,118.725 ns |   6.5303 ns |   5.4531 ns |  1,117.834 ns |
| Max (pr)          | 4096         |    607.103 ns |  11.9408 ns |  11.1694 ns |    600.681 ns |
| Max (main)        | 65536        | 31,133.123 ns | 452.4921 ns | 423.2614 ns | 31,015.117 ns |
| Max (pr)          | 65536        |  9,775.156 ns | 170.3118 ns | 159.3097 ns |  9,666.399 ns |
| Max (main)        | 131072       | 62,373.737 ns | 934.9628 ns | 874.5648 ns | 62,199.103 ns |
| Max (pr)          | 131072       | 19,485.664 ns | 372.0646 ns | 382.0830 ns | 19,188.719 ns |